🔧 FIX: Remove casas decimais desnecessárias das quantidades
✅ CORREÇÕES IMPLEMENTADAS:

📊 MODELOS CORRIGIDOS:
- Material.js: quantidade e quantidadeMinima → INTEGER
- Movimentacao.js: quantidade, quantidadeAnterior, quantidadeNova → INTEGER
- Requisicao.js: quantidade → INTEGER

🔧 ROTAS ATUALIZADAS:
- materials.js: parseFloat → parseInt
- movimentacoes.js: parseFloat → parseInt

🎨 FRONTEND CORRIGIDO:
- utils.js: formatQuantity sem casas decimais
- api.js: funções de movimentações
- Páginas de ajuste/estoque: salva histórico

🗄️ BANCO DE DADOS:
- Script de migração executado
- Campos convertidos para INTEGER
- Valores existentes convertidos para inteiros

📋 NOVAS FUNCIONALIDADES:
- Sistema completo de histórico de movimentações
- Página de histórico de produtos
- APIs para movimentações
- Modelo Movimentacao.js
- Rotas movimentacoes.js

🎯 RESULTADO:
- Quantidades salvas como números inteiros
- Sem casas decimais desnecessárias (.00)
- Interface limpa e intuitiva
- Histórico completo de movimentações

+++ server/routes/materials.js
@@ -24,7 +24,7 @@ router.get('/estoque/baixo', auth, async (req, res) => {
 
     // Filtrar materiais com estoque baixo
     const materiaisEstoqueBaixo = materiais.filter(material => 
-      parseFloat(material.quantidade) <= parseFloat(material.quantidadeMinima)
+      parseInt(material.quantidade) <= parseInt(material.quantidadeMinima)
     );
 
     res.json({ materiais: materiaisEstoqueBaixo });
@@ -183,8 +183,8 @@ router.post('/', auth, almoxarife, async (req, res) => {
     const material = await Material.create({
       nome,
       unidadeDeMedida,
-      quantidade: parseFloat(quantidade) || 0,
-      quantidadeMinima: parseFloat(quantidadeMinima) || 0,
+      quantidade: parseInt(quantidade) || 0,
+      quantidadeMinima: parseInt(quantidadeMinima) || 0,
       descricao,
       categoria,
       criadoPor: req.user.id
@@ -245,8 +245,8 @@ router.put('/:id', auth, almoxarife, async (req, res) => {
     // Atualizar campos
     if (nome) material.nome = nome;
     if (unidadeDeMedida) material.unidadeDeMedida = unidadeDeMedida;
-    if (quantidade !== undefined) material.quantidade = parseFloat(quantidade);
-    if (quantidadeMinima !== undefined) material.quantidadeMinima = parseFloat(quantidadeMinima);
+    if (quantidade !== undefined) material.quantidade = parseInt(quantidade);
+    if (quantidadeMinima !== undefined) material.quantidadeMinima = parseInt(quantidadeMinima);
     if (descricao !== undefined) material.descricao = descricao;
     if (categoria !== undefined) material.categoria = categoria;
 
@@ -335,7 +335,7 @@ router.post('/:id/adicionar-estoque', auth, almoxarife, async (req, res) => {
       });
     }
 
-    await material.adicionarEstoque(parseFloat(quantidade));
+    await material.adicionarEstoque(parseInt(quantidade));
 
     res.json({
       message: 'Estoque atualizado com sucesso',
